requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
openpyxl>=3.1.2
plotly>=5.18.0
fpdf2>=2.7.8
numpy>=1.26.0
scipy>=1.12.0
xlsxwriter>=3.1.9
